- Rename Node Provider interface to INodeProvider - Turn node-provider into a rollup package
