+++ packages/node-provider/src/index.ts
@@ -1,4 +1,4 @@
 import NodeProvider from "./node-provider";
 export { NodeProvider };
 
-export * from "./interfaces";
+export * from "./types";
